- created named queries in RawMaterialEntity; - created CRUD methods in RawMaterialDAO - created new functions in RawMaterialServices and controller

+++ src/main/java/org/internship/market/dto/RawMaterialDTO.java
@@ -14,6 +14,5 @@ public class RawMaterialDTO {
     private String name;
     private double price; // for 1 gram
     private double quantity;
-    private int stock;
-
+    private double stock;
 }
